Fix demo website not loading
I've used `locreq` to help resolving paths that are relative to the
root of the project, which should make the `assets` directory always
possible to properly locate on disk.

I've change the imports of package.json because the presence of a copy
of package.json messed up locreq's project root detection.